requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt
asgiref==3.8.1
drf-spectacular==0.28.0
daphne==4.1.2
adrf==0.1.9
